refactor(util): add new packages for util section
